Support for Mesos DNS API in aliPublish
Some hosts where aliPublish runs cannot be configured to use Mesos DNSes. This
feature adds a new configuration variable, `mesos_dns`, with a list of Mesos DNS
hosts. If a hostname in the configuration ends with .mesos, an attempt to
resolve it to IP using the [Mesos DNS API](https://mesosphere.github.io/mesos-dns/docs/http.html)
is made.