No fall back on dist when dist-runtime fails
A temporary connection failure might be the cause of dist-runtime unavailable.
In this case, considering all dist packages as runtime dependencies as fallback
is wrong and dangerous, as it could cause (as it happened) dozens of temporary
packages being unintentionally published. Now if the URL is not accessible (i.e.
if we cannot get the list of runtime dependencies for a given package) the
package is simply skipped. In case of a temporary server error a subsequent run
should recover.