Short-circuit consistency tests for PR (#401)
This should solve the infamous "PR check takes ~20 minutes" reports.
What was happening is that in order to ensure consistency of the builds
we triggered a full rebuild of development packages for each pull
request. This makes sense in the case of alidist pull requests because
you could have non-trivial interpackage dependencies like "use ROOT6
rather than ROOT5" which would fail miserably if short-circuited, at
least in the generic case. This is however not the case for AliRoot /
AliPhysics PR, where make / CMake takes care of the consistency for us
making the extra rebuild unneeded.