Fix the error log content
If no errors are found in a given log, we take a limited number
of last lines. Otherwise we put all the error lines found. Also,
we do not limit the cumulated total lines from all logs as we were
previously doing. This prevents accidentally cutting off errors from
certain logs (if there are many errors).