Pick a failed/reviewed PR to re-test randomly
Current C.I. mechanism favors untested PRs, and in case no PR is untested, it
retests the first one which failed (or the first one which was reviewed), not
giving any chance to other failed PRs to get retested. This PR mitigates the
problem by picking a random failed (or reviewed) PR instead of the first one.