Make coverage work for branches
This changes two things in continuous builder to make coverage
report also for branches.

First of all commits on branches in the main repository are always
considered "trusted" since someone must have merged them before.

Secondly, in the case of branches the PR number is actually the branch
name, so we do not pass that to codecov, since it seems to be confused
by it and drops the report.