Change naming convention for automatic RC (#568)
Plenty of hardcoded stuff in AliRoot/AliPhysics expects releases to resemble
something like v5-09-31. Portions of code are even loaded conditionally,
according to the AliRoot/AliPhysics version detected! To make all this legacy
machinery happy, we use the following naming convention for the release
candidates:

  * AliRoot: v9-99-99-rcYYYYMMDD
  * AliPhysics: v9-99-99-01-rcYYYYMMDD
  * AliDPG: v9-99-XX-rcYYYYMMDD

where the `v9-99-99` and `v9-99-XX` are literal, and `YYYYDDMM` is the current
date in the Geneva (CH) timezone. Using v9-99 should mean to all the conditional
parts that we are testing "the latest and greatest".