continuous-builder: allow updating continuous builder as it runs (O2-1366) (#749)
The most annoying bit of technical debt about continuous-builder.sh is that in
order to deploy some change one needs to restart the builder itself, cooling
the build cache and requiring a first round of packages to be built before we
can see the effects of some minor change, e.g.  sending a new metric about the
build process itself. This removes the problem by moving the inner loop in a
separate script which is sourced by the driver, effectively picking up changes
at the next iteration.