Unified CI builder (#801)
* Remove ONESHOT mode

This was previously intended to warm up the cache but that use case didn't
apply, and doesn't really apply to the new unified builders.

* Extract commonly used commands into helper functions

This adds the following functions:

- {short,long}_timeout to replace $TIMEOUT_CMD
- clean_env so sensitive environment variables are all cleared in one place
- get_config and get_config_value to extract repetitive code out and to allow
  reading the configuration more often, more easily
- reset_git_repository because I'd like to do that in two places
- report_pr_errors to reduce verbosity

* Reinitialise repo-dependent variables on every iteration

* Loop through repos

This is the heart of the PR. Here, we loop through all repositories that are
compatible with the docker image we are running under. The definitions for that
are stored in ci/repo-config/*/*.env.

* Make code more robust

This adds lots of quoting, gets rid of some variables storing exit codes (it now
just tests the commands themselves), fixes shellcheck issues and generally
improves the code style.

* Adjustable job role

Repositories to build are selected based on the MESOS_ROLE variable.

* Search parent directories for config/

As we cd into repo-specific directories, this allows us to set config/ variables
in a single place.

* Allow reinstalling ali-bot for running worker

This saves us having to restart workers when we change any scripts.

* Add a dedicated mode to only test already-tested PRs

This is to be used by a separate worker that will only test previously-tested
PRs so e.g. broken tests not due to bad code are eventually repaired.