Extract error messages into separate HTML file (#824)
* Extract repetitive code
* Extract error messages into separate HTML file

This extracts the important messages out of the large log files, so users can
find the more easily. The HTML file is uploaded to S3 alongside the full log
file, and links to the latter so both are accessible. The GitHub status will
link to the HTML file directly.

* Extract o2checkcode errors and test results separately

This deemphasises the old error list (but keeps it intact), and presents errors
from unit tests and from o2checkcode first.