Don't report errors to alidist (#837)
Not sure why this was there in the first place. Errors should always go to the
commit that was tested; PR_REPO and PR_HASH are always set, so no need for a
fallback. Setting the HEAD of alidist to pending on every random build messes up
PRs to that repo.