Speed up repo-s3-upload (#842)
* Speed up repo-s3-upload

This makes find read the link target directly, which is hopefully faster than
using readlink. The script is now designed to be run without set -x (which gives
far too much output).

* Search only new directories for symlinks

This should speed up symlink creation a lot, since disk reads in the repo are
throttled significantly. With this patch, only changed directories are searched.

With ~600k symlinks in ~40k directories and few changes from run to run, this
should theoretically increase run times by a factor of ~10.