Clean up config/ handling
We only need to check up to one directory level up, so avoid checking for
config/ directories in unrelated places on the system. This patch reduces the
number of log messages a bit, as well.

The patch also gets rid of the JOBS variable, which should be set in the
per-repo .env files. A global override doesn't really make sense as it needs to
be tweaked per-repo anyway.