Disable remote write store for PRs (#848)
* Disable remote write store for PRs

REMOTE_STORE is made read-only if we're building PRs. This makes sure only the
results of builds for the master branch are uploaded.

Development packages are still not uploaded in any case, as aliBuild handles
that by disabling the remote store.

* Enable RW store only for alidist checks

* Use HTTPS API to S3 when reading; S3 when writing

And make the MacOS alidist builder also write built packages, like the others.

* Use S3 remote only when testing branch

This allows for using the faster HTTPS API when building PRs, which should not
be published.

* Don't upload anything from slc7 yet

* Process all commits from master on the same worker

Master branches from different checks/repos can still be built on different
workers.