Fix spurious shlex error (#886)
This processes whole .env files at a time, instead of line-by-line.

Note that it doesn't parse files perfectly! A command with an argument
containing "=" would cause that argument to be parsed as an assignment. For
example, `echo foo=bar' would parse as an ignored token `echo' and an assignment
`foo=bar'. (This might be useful with export statements, actually.) As .env
files are supposed to be simple key-value stores anyway, we can ignore this
problem for now.