Run CI iterations in a subshell (#899)
We need this as we `cd' into repo-specific directories inside the loop. The
automatic subshell around the loop doesn't get us back to the sandbox directory
after each iteration.

This means we don't query for pending PRs after each iteration but only when
done iterating, as intended. This means new incoming PRs don't displace old
untested ones in the queue.