Push alidist tag to ALIDIST_BRANCH
If we get to the end of the script, the requested tag either doesn't exist in
the package's repo, or it's OK to overwrite. Therefore, we can force-push the
alidist tag as well.

If we can't directly push to ALIDIST_BRANCH, we instead push to an rc/ branch. A
GitHub action in the alidist repo then creates a PR from it.