Minor improvements to list-branch-pr (#925)
* Build older PRs first

For new (yellow) PRs, older ones have been in the queue longer and should this
be built first in the interest of fairness.

For previously tested (red and green) PRs, we take a single random PR to test,
so the sorting has no effect.

* Only fall through to green PRs if there are any

This modifies the check slightly so that if there are no green or yellow PRs, a
red one is always built. Previously, there was a 30% chance of falling through
the check for red PRs, even if there were no green ones, leading to a 10-minute
sleep.