Modify both defaults-%s.sh and %s.sh; build using branch before tagging (#910)
* Extract tag editing script

This is so that we can build using an rc branch but refer to the appropriate tag
in the release commit.

* Also add and commit package definition

In addition to defaults file. Both files are edited in edit_tags, added in the
previous commit.

* Push alidist tag to ALIDIST_BRANCH

If we get to the end of the script, the requested tag either doesn't exist in
the package's repo, or it's OK to overwrite. Therefore, we can force-push the
alidist tag as well.

If we can't directly push to ALIDIST_BRANCH, we instead push to an rc/ branch. A
GitHub action in the alidist repo then creates a PR from it.