Only increase ulimit if needed
The largest allowable limit seems to be 10240 on alibuildmac00. The others have
a maximum limit of 24576, and on Linux the maximum limit is even higher. This
patch doesn't change the ulimit if it is already large enough, as the O2 full
system test on Linux needs a very high ulimit.