daily-tags: Don't take up an slc7-light node when it's not needed
Currently, the Jenkins daily-tags script takes up a slot on an slc7-light node
for the full duration of the build, even though that slot is only needed for the
initial step of waiting for PRs to be merged (and most daily builds don't even
use that step).

With this patch, that slot is only used as long as it is needed, and is freed
when the real build begins on another node.

The diff contains mostly whitespace changes.