Add script to cleanup tarballs on S3
This script takes packages names, and searches for tarballs of them that are
older than a configurable cutoff. Old tarballs are deleted only if no other
tarball depends on them.

In addition to deleting the main tarballs, this script cleans up leftover
symlinks in `TARS/$arch/$package/` and in `dist*/` as well.