bulk-change-pr-status feature: Reset to previous status (#1055)
Add a feature to bulk-change-pr-status that allows changing the status of a
commit back to what it was before the latest status was set. This is better for
cleaning up after broken builders, as if things were green before, they can go
back to that, while PRs that didn't have a status before things were broken will
be yellow. This means that fixed builders have potentially much less of a queue
to rebuild.

A special case is that if the previous status is red, then we force-reset it to
pending because the logs from the earlier red status will have been overwritten
by now.