Use alibuild's docker support for CI (#1037)
* Use aliBuild's built-in Docker support

We still want to build in only one Docker image for each worker, so that the
local store doesn't get messed up. This makes it easy: we can get the
architecture and image name from existing job parameters.

* Use pip3 for CI

Plain pip isn't installed on the alibuild* machines.

* Auto-update and cleanup Docker images

* Use different option syntax for passing Docker args to aliBuild

Without the equals sign, aliBuild gets confused when the extra docker args start
with "--".

* Use slc* architecture for CentOS Stream

* Pass git credentials through to build container

These are needed to clone protected repositories, e.g. TpcFecUtils.

* Remove duplicate unneeded variables

These are now picked up from the Docker container, instead of having to be set
here.

`ALIBUILD_O2_TESTS` is an exception -- it is used for all tests (except
slc8-gpu), but isn't set in the containers.

* Don't use --upgrade-strategy only-if-needed

Some hosts have old versions of pip installed and don't support the option.

* Only use Docker on Linux systems

On MacOS, we still build on the bare system, so disable flags specific to builds
in Docker there.