daily tags: If no tag is set in recipe, add it
If the recipe we're building doesn't specify a `tag:` key, we need to add one,
as at this point in the script we cannot use the `version:` key instead (the
latter doesn't have an `rc/` prefix.)

This misses the case where no tag is set in the recipe, but the default's
override section sets a tag. In that case, a `tag:` line would be added to the
recipe, but would be overridden by the default.