Update check message when rebuilding
When we're automatically rebuilding a check, it would be helpful for users to
have feedback about this, so change the existing status's message (keeping the
red/green status and results URL intact).

This requires an extra option to set-github-status to keep the URL intact, which
is added by this commit as well.