Clean up unused builds after 2 days
After every build, delete builds and installations of development packages that
haven't been touched in two days. This should hopefully keep disk usage under
control.

A better solution would probably be to monitor the absolute disk usage of the CI
builder and delete as many old builds as we have to to stay within our limit,
but that would be much more complicated, and counting our own disk usage would
take a long time.

Also, monitor the time taken by the cleanup process. For this purpose, factor
out a function for pushing to InfluxDB, which could be used to push additional
metrics in the future.