Fix CI cleanup
This fixes a mistake in the invocation of the cleanup script, leading it to look
for files to clean up in the wrong places. This only occurred on builders where
`ALIBOT_CONFIG_SUFFIX` was set (those split off from the "main" builders
for their respective platforms), i.e. the O2Physics and AliPhysics builders.