Speed up cleanup significantly
Use symlink manifest files to find symlink targets, so we don't have to fetch
every symlink in dist*/ individually.

This is possibly slightly more confusing in case there are two tarballs with the
same version and revision number (in which case the nominal dependencies of one
may be deleted, even if the other tarball itself is not deleted), but in this
case aliBuild's and the publishers' behaviour is broken anyway, so never mind.

This improvement constitutes a ~6-10x speed-up for large repos.