Fix reuse of nightly tarballs
Since we now delete rc/* branches on successful builds, there is nothing for
later aliBuild invocations to compare against, so they cannot reuse tarballs
from nightly builds.

Instead, use the commit hash of the rc/ branch as the tag, so that the tarball's
hash is something aliBuild can discover later.