Run CI build only when specific paths have changed
In some cases, it would be good to run the CI only if some specific files have
changed (e.g. MC/ and tests/ in O2DPG). It may also be good to avoid running the
build if only documentation has changed (e.g. *.md, *.txt).

This is possible with the new `ONLY_RUN_WHEN_CHANGED` variable, which is in "git
diff" pathspec format. For example:

    # Only run CI if anything in MC/ and/or tests/ has changed.
    ONLY_RUN_WHEN_CHANGED='MC/ tests/'
    # Skip CI if only documentation has changed.
    ONLY_RUN_WHEN_CHANGED=':!\*.txt :!\*.md'

If the CI is skipped, the build is marked as successful, to avoid blocking
merges if checks are marked as required, and because the GitHub status API does
not allow marking statuses as skipped.