Track commit push time
This should give a good approximation of the time a commit was submitted for
testing. It still breaks down for draft PRs that are marked non-draft after a
while, but PRs that are opened as non-draft from old commits are handled
correctly.

The alternative would be to use the PR's updatedAt property, but then any
comments posted in the PR would reset the waiting timer.