Use setuptools_scm for automatic versioning (#1245)
This is the same approach that alibuild already uses.

Recent pip versions have stopped accepting "LAST_TAG" as a valid
version number. We used a constant "0.0.0" version as a stop-gap, but
this still had problems -- pip didn't update the installed ali-bot
version sometimes, since it assumed that the right version was already
installed.

This new, proper versioning scheme encodes the commit number (and
hash) in the autogenerated version number, so auto-updates (e.g. in
CI) should work properly.

This will allow us to get rid of some hacky workarounds in the CI
scripts as well.

Tested and working on CentOS 7 and above under Python 3.6+.