Update clang-format to v15 to match alidist
It would be less potentially confusing if we ran the same
git-clang-format version in CI as people will run locally. Locally, we
expect people to run git-clang-format in alienv, so they'll get the
one installed with alidist, which is currently v15.

Clang-format v15 is not preinstalled in any GitHub runner images, so
we have to explicitly install it, but we can also get rid of the
"ubuntu-22.04" pinning.