Auto-detect assigned number of cores for Nomad jobs
This allows Jenkins builders to detect how many cores they have been assigned
by Nomad, so that we don't have to hard-code JOBS variables in Jenkins. Fall
back to the total number of cores if this detection doesn't work, since if the
Jenkins builder isn't running under Nomad, it's likely the only thing running
on the machine.