Use all cores allocated to Jenkins builder by default
Our Nomad job now exposes a MAX_CORES variable.

Always use JOBS if given, but fall back to MAX_CORES and then $(nproc) if that
variable is not given.

This should allow transparently using 32 cores on the new, larger builders,
and solve the OOM errors on ARM when we use too many jobs by default.