Build QC as CI for alibuild pull requests
This installs the alibuild version with changes proposed in the PR, then
builds QC with that alibuild version to test it.

This is fine to run on the same build machines as other CI, since we isolate
the Python environment in a venv and the next build will overwrite alibuild
with its own version.