Extend symlink caching script to handle dist symlinks (#1248)
* Extend symlink caching script to handle dist symlinks

This should make publishing much faster, once publishers can start to
use pre-built manifests to avoid having to walk the full dist tree on
S3 themselves.

* Fix bytes handling

* Fix architecture matching

This makes globs like "slc7_*" work as parameters for -a.

* Fix quitting when some workers fail

* Save S3 calls by relying on caches more

This uses information cached from previous runs. It has the danger that it
treats this information as authoritative, so if "real" symlinks' targets have
changed in the meantime, this will not catch it, but it speeds this script up
significantly, especially for dist trees.

* Trace put_object calls

* Skip uploads if nothing has changed

* Trace last remaining untraced S3 API call

* Improve logging

* Abort on the first missing tarball

This avoids having to read the rest of the symlinks in the directory, reducing
the number of read_object calls for slc7 from 4488 to 927 on a run with no
changes.